Fix pipeline for acceptance environment

+++ scripts/git_commons.py
@@ -1,11 +1,11 @@
 import os
 import logging
-import git
+from git import Repo, InvalidGitRepositoryError
 
 logger = logging.getLogger(__name__)
 
 
-def get_git_repo(repo_dir):
+def get_git_repo(repo_dir: str) -> Repo:
     """
     Returns the git repository used for VCS of source and transformed data files. As well as load logs.
     If it does not exist, it will create one.
@@ -14,26 +14,32 @@ def get_git_repo(repo_dir):
     """
     if not os.path.exists(repo_dir):
         return init_git_repo(repo_dir)
-
     try:
-        return git.Repo(repo_dir)
-    except git.InvalidGitRepositoryError:
+        repo = Repo(repo_dir)
+        init_gitignore(repo)
+        return repo
+    except InvalidGitRepositoryError:
         return init_git_repo(repo_dir)
 
 
-def init_git_repo(repo_dir):
-    os.makedirs(repo_dir, exist_ok=True)
-    logger.info('Initializing git repository: {}'.format(repo_dir))
-    r = git.Repo.init(os.path.realpath(repo_dir))
+def init_gitignore(repo: Repo):
+    gitignore = os.path.realpath(os.path.join(repo.working_tree_dir, '.gitignore'))
+    if os.path.isfile(gitignore):
+        return
+
     ignore_list = ['.done-*', '.DS_Store']
-    logger.debug('Git ignore list: {}'.format(ignore_list))
-
-    gitignore = os.path.realpath(os.path.join(repo_dir, '.gitignore'))
-
     logger.debug('Writing git ignore file')
     with open(gitignore, 'w') as f:
         f.write('\n'.join(ignore_list))
+        f.write('\n')
 
-    r.index.add([gitignore])
-    r.index.commit('Initial commit.')
-    return r
+    repo.index.add([gitignore])
+    repo.index.commit('Initial commit.')
+
+
+def init_git_repo(repo_dir) -> Repo:
+    os.makedirs(repo_dir, exist_ok=True)
+    logger.info('Initializing git repository: {}'.format(repo_dir))
+    repo = Repo.init(os.path.realpath(repo_dir))
+    init_gitignore(repo)
+    return repo
